Moving DNS/IP resolving to separate class in Scrapy module.
